requirements.txt
fastapi
uvicorn
google-generativeai
chromadb
kaggle
slowapi
